refactor: make DanmakuFactory protected & package-local
